feat : 대기열 토큰 대기번호 로직 변경 #29 - 대기열 토큰 생성할때 대기번호 부여하는 방식에서 조회할때 부여하는 방식으로 변경

+++ src/main/java/com/kpop/ticketing/presentation/waittoken/dto/response/WaitTokenResponse.java
@@ -10,15 +10,15 @@ import lombok.NoArgsConstructor;
 public class WaitTokenResponse {
 	private String token;
 	private WaitingStatus waitingStatus;
-	private Integer waitingNumber;
+	private long waitingNumber;
 
-	private WaitTokenResponse(String token, WaitingStatus waitingStatus, Integer waitingNumber) {
+	private WaitTokenResponse(String token, WaitingStatus waitingStatus, long waitingNumber) {
 		this.token = token;
 		this.waitingStatus = waitingStatus;
 		this.waitingNumber = waitingNumber;
 	}
 
-	public static WaitTokenResponse of(String token, WaitingStatus waitingStatus, Integer waitingNumber) {
+	public static WaitTokenResponse of(String token, WaitingStatus waitingStatus, long waitingNumber) {
 		return new WaitTokenResponse(token, waitingStatus, waitingNumber);
 	}
 }
